requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
audioplayer==0.6
autopep8==1.5.4
aws==0.2.5
awsebcli==3.20.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
boto==2.49.0
boto3==1.17.59
botocore==1.20.59
bs4==0.0.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
comtypes==1.1.8
cryptography==3.4.7
disjoint-set==0.7.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.5
django-brutebuster2==0.2.0
django-ckeditor==6.0.0
django-common-helpers==0.9.2
django-js-asset==1.2.2
django-storages==1.11.1
fabric==2.6.0
filelock==3.0.12
flake8==3.8.4
future==0.16.0
gTTS==2.2.1
gunicorn==20.1.0
idna==2.10
invoke==1.5.0
isort==5.6.4
jedi==0.17.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mpg123==0.4
mysqlclient==2.0.3
paramiko==2.7.2
parso==0.7.1
pathlib2==2.3.5
pathspec==0.5.9
paypalrestsdk==1.13.1
Pillow==8.2.0
playsound==1.2.2
pluggy==0.13.1
prettytable==2.1.0
psycopg2==2.9.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pydub==0.24.1
pyflakes==2.2.0
pygame==2.0.1
pyglet==1.5.14
pylint==2.6.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.5
PyYAML==5.4.1
ratelimit==2.2.1
requests==2.25.1
rope==0.18.0
s3transfer==0.4.2
semantic-version==2.8.5
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.1
sqlparse==0.4.1
stripe==2.55.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.55.0
ujson==4.0.1
urllib3==1.25.11
virtualenv==20.2.2
wcwidth==0.1.9
wget==3.2
whitenoise==5.3.0
wrapt==1.12.1
yapf==0.30.0
